requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
amqp==5.2.0
attrs==23.2.0
billiard==4.2.0
celery==5.4.0
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
discord.py==2.4.0
frozenlist==1.4.1
idna==3.7
kombu==5.3.7
multidict==6.0.5
numpy==2.0.0
pandas==2.2.2
prompt_toolkit==3.0.47
pybit==5.7.0
pycryptodome==3.20.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
requests==2.32.3
schedule==1.2.2
six==1.16.0
tzdata==2024.1
urllib3==2.2.2
vine==5.1.0
wcwidth==0.2.13
websocket-client==1.8.0
websockets==12.0
yarl==1.9.4
